// src/features/dashboard/types.ts
export interface DashboardSummary {
    totalEquipment: number;
  
    availableEquipment: number;
  
    assignedEquipment: number;
  
    maintenanceEquipment: number;
  
    activeAssignments: number;
  
    completedAssignments: number;
  
    activeRentals: number;
  
    returnedRentals: number;
  
    overdueRentals: number;
  
    scheduledMaintenance: number;
  
    maintenanceInProgress: number;
  
    completedMaintenance: number;
  
    upcomingReturns: number;
  }